+++ CleanArchitecture/Application/Features/Categories/Queries/CategoryQueryHandler.cs
@@ -8,13 +8,13 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace Application.Features.Categories.Queries.FetchById
+namespace Application.Features.Categories.Queries
 {
-    public class GetByIdCategoryHandler : IRequestHandler<GetByIdCategoryQuery, CategoryDTO>
+    public class CategoryQueryHandler : IRequestHandler<GetByIdCategoryQuery, CategoryDTO>,IRequestHandler<GetAllCategoriesQuery,List<CategoryDTO>>
     {
         private readonly IUnitOfWork unitOfWork;
         private readonly IMapper mapper;
-        public GetByIdCategoryHandler(IUnitOfWork unitOfWork,IMapper mapper)
+        public CategoryQueryHandler(IUnitOfWork unitOfWork,IMapper mapper)
         {
             this.unitOfWork = unitOfWork;
             this.mapper = mapper;
@@ -25,7 +25,11 @@ namespace Application.Features.Categories.Queries.FetchById
             var category = await unitOfWork.Categories.GetByIdAsync(request.Id);
             return mapper.Map<CategoryDTO>(category);
         }
+        public async Task<List<CategoryDTO>> Handle(GetAllCategoriesQuery request, CancellationToken cancellationToken)
+        {
+            var categories= await unitOfWork.Categories.GetAllAsync();
+            return mapper.Map<List<CategoryDTO>>(categories);
+        }
 
-       
     }
 }
