--- CleanArchitecture/Application/Features/Categories/Queries/FetchAll/GetAllCategoriesHandler.cs
@@ -1,30 +1,0 @@
-﻿using Application.DTOs;
-using AutoMapper;
-using Domain.Entities;
-using Domain.Generic;
-using MediatR;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Application.Features.Categories.Queries.FetchAll
-{
-    public class GetAllCategoriesHandler : IRequestHandler<GetAllCategoriesQuery, List<CategoryDTO>>
-    {
-        private readonly IUnitOfWork unitOfWork;
-        private readonly IMapper mapper;
-        public GetAllCategoriesHandler(IUnitOfWork unitOfWork,IMapper mapper)
-        {
-            this.unitOfWork = unitOfWork;
-            this.mapper = mapper;
-        }
-        public async Task<List<CategoryDTO>> Handle(GetAllCategoriesQuery request, CancellationToken cancellationToken)
-        {
-            var categories =await unitOfWork.Categories.GetAllAsync();
-            var categoriesDTOs = mapper.Map<List<CategoryDTO>>(categories);
-            return categoriesDTOs;
-        }
-    }
-}
